extract async_test decorator
